requirements.txt
APScheduler==3.6.3
wheel
Django
gunicorn
psycopg2
whitenoise
django-heroku
python-telegram-bot